src/api/requirements.txt
azure-functions
fastapi
pydantic
